changes made while reading the code to identify the rss mem usage

+++ utils/mem_mgr.h
@@ -32,7 +32,7 @@ class MemoryManager {
   uint64_t chunk_size_;
 
   // Total number of chunks in this allocator.
-  int num_chunks_;
+  size_t num_chunks_;
 
   // A list of free buffers
   std::list<uint8_t*> free_buffers_;
